stylo_derive

Module to_typed

Structs

TypedFieldAttrs

TypedInputAttrs

TypedVariantAttrs

Functions

derive

Derive implementation of the ToTyped trait.

derive_bitflags 🔒

Derive a ToTyped implementation for CSS bitflags types.

derive_single_field_expr 🔒

Generate the expression used to reify a single field in a derived ToTyped implementation.

derive_variant_arm 🔒

Generate the match arm expression for a struct or enum variant in a derived ToTyped implementation.

derive_variant_fields_expr 🔒

Generate the match arm expression for fields of a struct or enum variant in a derived ToTyped implementation.

field_generic_arguments 🔒

Extract generic type arguments from a field type.