stylo_derive::to_typed

Struct TypedValueVariantAttrs

pub struct TypedValueVariantAttrs {
    pub derive_fields: bool,
    pub skip: bool,
    pub todo: bool,
}

Fields
derive_fields: bool

Same as the top-level derive_fields, but included here because struct variants are represented as both a variant and a type definition.

skip: bool

If present, this variant is excluded from generated reification code. to_typed() will always return None for it.

todo: bool

Marks this variant as a placeholder for a future implementation. Behavior is the same as skip, but used to indicate that reification is intentionally left unimplemented for now.