stylo_derive::to_css

Struct CssBitflagAttrs

pub struct CssBitflagAttrs {
    pub single: Option<String>,
    pub mixed: Option<String>,
    pub validate_mixed: Option<Path>,
    pub overlapping_bits: bool,
}

Fields
single: Option<String>

Flags that can only go on their own, comma-separated.

mixed: Option<String>

Flags that can go mixed with each other, comma-separated.

validate_mixed: Option<Path>

Extra validation of the resulting mixed flags.

overlapping_bits: bool

Whether there are overlapping bits we need to take care of when serializing.

Implementations

impl CssBitflagAttrs

fn names(s: &Option<String>) -> Vec<(String, String)>

Returns a vector of (rust_name, css_name) of a given flag list.

pub fn single_flags(&self) -> Vec<(String, String)>

pub fn mixed_flags(&self) -> Vec<(String, String)>